Introduce zone id and mapping to matrix column
Use zone id instead of oid.

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/populationsynthesis/ipu/DefaultHouseholdBuilder.java
@@ -35,7 +35,7 @@ public class DefaultHouseholdBuilder implements HouseholdBuilder {
 		currentHousehold = householdCreator.createHousehold(panelHousehold, zone());
 		createSimulatedPeople(panelHousehold);
 		demandZone.actualDemography().incrementHousehold(currentHousehold.getSize());
-		int income = (int) currentHousehold.monthlyIncomeEur();
+		int income = currentHousehold.monthlyIncomeEur();
     demandZone.actualDemography().increment(StandardAttribute.income, income);
 		return currentHousehold;
 	}
